services/auth/requirements.txt
# Web Framework
fastapi==0.110.0
uvicorn[standard]==0.29.0

# Base de Datos y ORM
sqlalchemy==2.0.29
psycopg2-binary==2.9.9

# Seguridad y JWT
bcrypt==4.0.1
passlib[bcrypt]==1.7.4

# Validación y Configuración
pydantic[email]==2.6.4
python-dotenv==1.0.1

# Cliente HTTP (para comunicación entre microservicios futura)
httpx==0.27.0

python-jose[cryptography]